requirements.txt
django==2.0.7
wagtail
psycopg2-binary
djangorestframework